grow NFS request cache for the server side from 64 to 2048 entries.
FreeBSD did the corresponding change a while ago - the server request
cache is used to avoid giving unnecessary failures to clients when a request
is completed server side, but the client does not receive the result
and therefore retries the request.  This avoids errors like "file
already exists" when the client tries to create files on a busy
NFS server.
ok krw@, tedu@, dcoppa@, deraadt@