Do not trigger end-of-sentence spacing by trailing punctuation at the end of partial implicit macros.  Prodded by jmc@.
Actually, this is a revert of rev. 1.39 Sat May 15 09:20:01 2010 UTC,
which in turn is mdocml.bsd.lv rev. 1.64 Fri May 14 14:09:13 2010 UTC
by kristaps@, with this original commit message:
"Block-implicit macros now up-propogate end-of-sentence spacing.
NOTE: GROFF IS NOT SMART ENOUGH TO DO THIS."
Please speak after me: Then why the hell should we?

We already weakened this in rev. 1.55 Fri Jul 16 00:34:33 2010 UTC,
but that weakening was insufficient.  Let's take it out completely.

Admittedly, there are two places in base where what Kristaps did
made the output nicer, in calloc(3) and in fish(6).  But both are
atypical.  There are 18 other places where this revert makes the
output nicer, the typical case being:
"Mail status is shown as ``No Mail.''  if there is no mail."
You do *not* want the EOS spacing after ``No Mail.'' in that sentence.