Properly initialize secondary CPUs on 64bit machines.
Due to a bug in our Openfirmware client interface on G5, the call to
OF_finddevice() in cpu_spinup() will never return if the given argument
is on the stack.  Use a workaround for the moment and put it in the bss.

Yes, this is ugly, but it makes GENERIC.MP ``usable'' on Powermac7,2 and
7,3 and does not freeze your machine during the first boot.

Issue also reported by "The Mouse" on ppc@, tested by landry@