Mark Cave-Ayland pointed out on tech@ that the OF_test_method was called with an ihandle instead of a phandle. And also the comparison was the wrong way around.
This bug was causing an exception. Because -1 is returned to indicate
failure and because the comparison is the wrong way around, the bad logic
led to believing that SUNW,retain exists which is why this currently
works on some PROMs.

On E250/E450 machines, this didn't work so there was a hack to detect
these machines and work-around the problem. After this fix that hack is
now removed.

Tested on an E450 by sebastia@ and confirmed that his dmesg is now
retained after a reboot. nick@ promises to complain if his E250 is any
worse off.

ok kettenis@